--- a/requirements/environment.yml
+++ b/requirements/environment.yml
@@ -12,7 +12,7 @@ dependencies:
   - click
   - cerberus
   - dask
-  - dessinemoi
+  - dessinemoi>=22.2.0
   - environ-config
   - iapws
   - matplotlib