--- a/requirements/main.in
+++ b/requirements/main.in
@@ -6,7 +6,7 @@ attrs
 click
 cerberus
 dask
-dessinemoi
+dessinemoi>=22.2.0
 environ-config
 iapws
 matplotlib