--- a/src/eradiate/radprops/_core.py
+++ b/src/eradiate/radprops/_core.py
@@ -18,6 +18,14 @@ from ..exceptions import UnsupportedModeError
 from ..units import unit_registry as ureg
 
 rad_profile_factory = Factory()
+rad_profile_factory.register_lazy_batch(
+    [
+        ("_afgl1986.AFGL1986RadProfile", "afgl_1986", {}),
+        ("_array.ArrayRadProfile", "array", {}),
+        ("_us76_approx.US76ApproxRadProfile", "us76_approx", {}),
+    ],
+    cls_prefix="eradiate.radprops",
+)
 
 
 @ureg.wraps(