--- a/src/eradiate/scenes/__init__.py
+++ b/src/eradiate/scenes/__init__.py
@@ -1,27 +1,20 @@
-from . import (
-    atmosphere,
-    biosphere,
-    bsdfs,
-    core,
-    illumination,
-    integrators,
-    measure,
-    phase,
-    shapes,
-    spectra,
-    surface,
+from ..util import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach(
+    __name__,
+    submodules=[
+        "atmosphere",
+        "biosphere",
+        "bsdfs",
+        "core",
+        "illumination",
+        "integrators",
+        "measure",
+        "phase",
+        "shapes",
+        "spectra",
+        "surface",
+    ],
 )
 
-__all__ = [
-    "atmosphere",
-    "biosphere",
-    "bsdfs",
-    "core",
-    "illumination",
-    "integrators",
-    "measure",
-    "phase",
-    "shapes",
-    "spectra",
-    "surface",
-]
+del lazy_loader