--- a/src/eradiate/scenes/atmosphere/__init__.py
+++ b/src/eradiate/scenes/atmosphere/__init__.py
@@ -1,41 +1,30 @@
-from ._core import (
-    AbstractHeterogeneousAtmosphere,
-    Atmosphere,
-    AtmosphereGeometry,
-    PlaneParallelGeometry,
-    SphericalShellGeometry,
-    atmosphere_factory,
-)
-from ._heterogeneous import HeterogeneousAtmosphere
-from ._homogeneous import HomogeneousAtmosphere
-from ._molecular_atmosphere import MolecularAtmosphere
-from ._particle_dist import (
-    ArrayParticleDistribution,
-    ExponentialParticleDistribution,
-    GaussianParticleDistribution,
-    InterpolatorParticleDistribution,
-    ParticleDistribution,
-    UniformParticleDistribution,
-    particle_distribution_factory,
-)
-from ._particle_layer import ParticleLayer
+from ...util import lazy_loader
 
-__all__ = [
-    "atmosphere_factory",
-    "particle_distribution_factory",
-    "AbstractHeterogeneousAtmosphere",
-    "ArrayParticleDistribution",
-    "Atmosphere",
-    "AtmosphereGeometry",
-    "ExponentialParticleDistribution",
-    "GaussianParticleDistribution",
-    "HeterogeneousAtmosphere",
-    "HomogeneousAtmosphere",
-    "InterpolatorParticleDistribution",
-    "MolecularAtmosphere",
-    "ParticleDistribution",
-    "ParticleLayer",
-    "PlaneParallelGeometry",
-    "SphericalShellGeometry",
-    "UniformParticleDistribution",
-]
+__getattr__, __dir__, __all__ = lazy_loader.attach(
+    __name__,
+    submod_attrs={
+        "_core": [
+            "AbstractHeterogeneousAtmosphere",
+            "Atmosphere",
+            "AtmosphereGeometry",
+            "PlaneParallelGeometry",
+            "SphericalShellGeometry",
+            "atmosphere_factory",
+        ],
+        "_heterogeneous": ["HeterogeneousAtmosphere"],
+        "_homogeneous": ["HomogeneousAtmosphere"],
+        "_molecular_atmosphere": ["MolecularAtmosphere"],
+        "_particle_dist": [
+            "ArrayParticleDistribution",
+            "ExponentialParticleDistribution",
+            "GaussianParticleDistribution",
+            "InterpolatorParticleDistribution",
+            "ParticleDistribution",
+            "UniformParticleDistribution",
+            "particle_distribution_factory",
+        ],
+        "_particle_layer": ["ParticleLayer"],
+    },
+)
+
+del lazy_loader