… CHANGELOG.md
… pyproject.toml
… requirements/dev.txt
… requirements/docs.txt
… requirements/environment-linux-64.lock
… requirements/environment-osx-64.lock
… requirements/main.txt
… requirements/tests.txt
… src/eradiate/_preimport.py
… src/eradiate/scenes/bsdfs/_black.py
… src/eradiate/scenes/bsdfs/_checkerboard.py
… src/eradiate/scenes/bsdfs/_core.py
… src/eradiate/scenes/bsdfs/_lambertian.py
… src/eradiate/scenes/bsdfs/_rpv.py
… src/eradiate/scenes/illumination/__init__.py
… src/eradiate/scenes/illumination/_constant.py
… src/eradiate/scenes/illumination/_core.py
… src/eradiate/scenes/illumination/_directional.py
… src/eradiate/scenes/integrators/__init__.py
… src/eradiate/scenes/integrators/_core.py
… src/eradiate/scenes/integrators/_path_tracers.py
… src/eradiate/scenes/measure/__init__.py
… src/eradiate/scenes/measure/_core.py
… src/eradiate/scenes/measure/_distant_flux.py
… src/eradiate/scenes/measure/_hemispherical_distant.py
… src/eradiate/scenes/measure/_multi_distant.py
… src/eradiate/scenes/measure/_multi_radiancemeter.py
… src/eradiate/scenes/measure/_perspective.py
… src/eradiate/scenes/measure/_radiancemeter.py
… src/eradiate/scenes/phase/__init__.py