# Description

This PR introduces lazy module loading to save time upon import so that using Eradiate interactively feels more comfortable. Changes are as follows:

* A lazy loading system is introduced in the `eradiate.util.lazy_loader` module, based on a [Scientific Python proposal](https://github.com/scientific-python/lazy_loader). It effectively replaces imports used to build flat namespaces with lazy loading capabilities. For this, it results that `import foo` only imports `foo` but makes `foo.bar` available immediately; `foo.bar` will however only be imported if `foo.bar.Baz` is requested.
* The *Dessine-moi* dependency is updated to v22.2+. It introduces lazy type registration and a convenient alias declaration system.
* The entire codebase was adapted to use this system. Lazy type registration to factories is required to make sure that factories can be used even if their registered types are not imported. A side effect is that factory registration code is now decoupled from classes themselves.
* Lock files are updated. Support of xarray 2022.06 required a small update to the `us76.make_profile()` function.

NB: This will likely disrupt IDEs: no stub files exist and they likely cannot resolve all names if not imported. However I think this tradeoff is worthwhile given how much more responsive `import eradiate` feels now. I expect IDE issues to be really problematic only if Eradiate is used as a library for script writing, *i.e.* not in an interactive console.

Example of performance / quality-of-life improvement (on AMD Ryzen 9; lighter configurations benefit more from this):

* **Before**
  ```
  ❯ time eradiate show
  eradiate show  1.97s user 0.45s system 117% cpu 2.070 total
  ```

* **After**
  ```
  ❯ time eradiate show
  eradiate show  0.89s user 0.28s system 150% cpu 0.785 total
  ```

# To do

- [x] Investigate bug introduced by changes to the `Dataset.rename_dims()` method in [xarray 2022.06](https://docs.xarray.dev/en/stable/whats-new.html?highlight=rename_dims#breaking-changes). This breaks the `us76.make_profile()` function.

# Checklist

- [x] The code follows the relevant coding guidelines
- [x] The code generates no new warnings
- [x] The code is appropriately documented
- [x] The code is tested to prove its function
- [x] The feature branch is rebased on the current state of the `main` branch
- [x] I updated the change log if relevant
- [x] I give permission that the Eradiate project may redistribute my contributions under the terms of its license